+++ week-04/day-02/src/Tree.java
@@ -5,7 +5,7 @@ public class Tree extends Plants{
   }
 
   @Override
-  public void watering(int wateramount) {
+  public void watering(double wateramount) {
     this.waterlevel = this.waterlevel + (wateramount * 0.40);
   }
 }
